Keep browser key tests focused
The latest review asked to separate auth sign-in coverage from browser key forwarding coverage. Moving the auth suite into its own test file keeps each test file under the size guidance without changing test behavior.

Constraint: Xcode project requires explicit test source membership

Rejected: Rename the existing browser test file | larger project-file churn without improving the requested split

Confidence: high

Scope-risk: narrow

Reversibility: clean

Directive: Keep auth callback coverage in AuthManagerBrowserSignInTests and key-routing coverage in BrowserArrowKeyForwardingTests

Tested: ./scripts/test-unit.sh build-for-testing -derivedDataPath /tmp/cmux-issue-3840-signin-beep-on-enter-unit

Tested: ./scripts/reload.sh --tag issue-3840-signin-beep-on-enter --launch

Not-tested: Runtime XCTest execution, per repository policy
